- Added start time prompt in Showtime creation - Added end time prompt in Showtime creation - Added end time calculation based on start time and duration of movie - Corrected 3 layer model in my code (may still be incorrect, check?)
