+++ Logic/Interfaces/IShowtimeService.cs
@@ -6,9 +6,8 @@ public interface IShowtimeService
 {
     Showtime? Find(int id);
     IEnumerable<Showtime> GetShowtimesByMovieId(int movieId, int limit = 10);
-    bool IsMovieIDValid(string movie);
-    bool CheckIfDataCorrect(string movie, int auditorium);
     void CreateShowtime(Showtime showtime);
-    int ShowMenuMovies(string title, List<Movie> options);
-    int ShowMenuAuditoriums(string title, List<Auditorium> options);
+    public bool IsMovieIDValid(string movie);
+    public bool IsMovieStartTimeValid(string time);
+    public DateTime parsedStartTime { get; set; }
 }
